platfrom web/backend/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
Flask-Caching==2.1.0

# Database
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# API and serialization
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.30.0

# Authentication and security
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0.post1

# File handling
Pillow==10.1.0
python-magic==0.4.27

# AI and data processing
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2

# WebSockets (for live features)
flask-socketio==5.3.6

# Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0 